Continued the ParserState implementation

+++ channel/channel.go
@@ -73,33 +73,33 @@ func NewLoopStack() *LoopStack {
 
 
 type Channel struct {
-    Name string             	// The channel's name ("A", "B", "C", etc)
-    Num int                 	// The channel's number (0, 1, 2, etc)
-    Ticks int			// The total number of ticks (32nd notes) used by this channel
-    Frames float64          	// The total number of frames used by this channel given the current refresh rate
+    Name string                 // The channel's name ("A", "B", "C", etc)
+    Num int                     // The channel's number (0, 1, 2, etc)
+    Ticks int           // The total number of ticks (32nd notes) used by this channel
+    Frames float64              // The total number of frames used by this channel given the current refresh rate
     LoopFrames float64
     LoopPoint int
     LastSetLength float64
-    CurrentTempo int        	// The channel's tempo, in BPM
+    CurrentTempo int            // The channel's tempo, in BPM
     CurrentNote Note
-    CurrentOctave int       	// The currently set octave for this channel
-    CurrentVolume int       	// The currently set volume for this channel
+    CurrentOctave int           // The currently set octave for this channel
+    CurrentVolume int           // The currently set volume for this channel
     Tuple struct {
         Cmds []Note
         HasData bool
         Active bool
     }
-    CurrentLength float64   	// Length in 32nd notes
+    CurrentLength float64       // Length in 32nd notes
     CurrentNoteFrames struct {
         Active, Cutoff float64
     }
     CurrentCutoff struct {
         Typ int
-        Val int			// Length in 32nd notes
+        Val int         // Length in 32nd notes
     }
     PendingOctChange int
-    HasAnyNote bool         	// Whether any notes have been added on this channel
-    Active bool             	// Is this channel currently active?
+    HasAnyNote bool             // Whether any notes have been added on this channel
+    Active bool                 // Is this channel currently active?
     Cmds []int
     UsesEffect map[string]bool
     Loops *LoopStack
@@ -262,9 +262,9 @@ func (chn *Channel) NoteLength(len float64) (frames, cutoffFrames, scaling float
         
         if (chn.CurrentCutoff.Typ == defs.CT_FRAMES ||
             chn.CurrentCutoff.Typ == defs.CT_NEG_FRAMES) {
-            cutoffFrames = math.Min(chn.CurrentCutoff.Val * scaling, frames)
+            cutoffFrames = math.Min(float64(chn.CurrentCutoff.Val) * scaling, frames)
         } else {
-            cutoffFrames = (frames * (8 - chn.CurrentCutoff.Val)) / 8
+            cutoffFrames = (frames * (8 - float64(chn.CurrentCutoff.Val))) / 8
         }
         frames = math.Floor(frames) - math.Floor(cutoffFrames)
         if (frames < 0) {
@@ -276,9 +276,9 @@ func (chn *Channel) NoteLength(len float64) (frames, cutoffFrames, scaling float
         frames = float64(length32) * len
         if (chn.CurrentCutoff.Typ == defs.CT_FRAMES ||
             chn.CurrentCutoff.Typ == defs.CT_NEG_FRAMES) {
-            cutoffFrames = math.Min(chn.CurrentCutoff.Val, frames)
+            cutoffFrames = math.Min(float64(chn.CurrentCutoff.Val), frames)
         } else {
-            cutoffFrames = math.Floor((frames * (8 - chn.CurrentCutoff.Val)) / 8)
+            cutoffFrames = math.Floor((frames * (8 - float64(chn.CurrentCutoff.Val))) / 8)
         }
         frames -= cutoffFrames
     }
@@ -316,7 +316,7 @@ func (chn *Channel) WriteLength() {
         }
         
     } else {
-	// ToDo: necessary to handle this?
+    // ToDo: necessary to handle this?
     }
 }
